Scrap srcds-server for steamcmd-server
The last project was confusing. Srcds refers to SOURCE servers, not steamcmd servers like the goal was. Furthermore, the code focused on SOURCE servers, and therefore failed for non-source steamcmd servers. From now on I base server images off of a steamcmd-server image that's based on a steamcmd image rather than an srcds-server image that has steamcmd embedded in it. These changes should make things a LOT clearer going forward.