Ability: share files with other servers
This will allow servers to share files. The shared files must be mounted in the servers' /shared and must have the exact same file structure as the servers. The primary reasoning for this is to allow sharing common files like maps without having to duplicate them and waste file space.